server integration test succeeded
