+++ server/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"fmt"
 	"log"
 	"net/http"
@@ -11,21 +12,28 @@ import (
 
 	"github.com/Oscar-W-Chen/gpu-ai-inference-server/inference_engine/binding"
 	"github.com/gin-gonic/gin"
+	"golang.ngrok.com/ngrok"
+	"golang.ngrok.com/ngrok/config"
 )
 
 func main() {
+	if err := run(context.Background()); err != nil {
+		log.Fatal(err)
+	}
+}
+
+func run(ctx context.Context) error {
 	// Initialize logger
 	log.SetFlags(log.LstdFlags | log.Lshortfile)
 	log.Println("Starting AI Inference Server...")
 
-	// Check CUDA availability directly from binding
+	// Check CUDA availability
 	cudaAvailable := binding.IsCUDAAvailable()
 	deviceCount := binding.GetDeviceCount()
 
 	log.Printf("CUDA Available: %v", cudaAvailable)
 	log.Printf("GPU Device Count: %v", deviceCount)
 
-	// Print device info for each GPU
 	if cudaAvailable && deviceCount > 0 {
 		for i := 0; i < deviceCount; i++ {
 			deviceInfo := binding.GetDeviceInfo(i)
@@ -35,16 +43,13 @@ func main() {
 
 	// Initialize router
 	router := gin.Default()
-
-	// Define simple routes
-	router.GET("/health", func(c *gin.Context) {
+		router.GET("/health", func(c *gin.Context) {
 		c.JSON(http.StatusOK, gin.H{
 			"status": "healthy",
 			"time":   time.Now().Unix(),
 		})
 	})
 
-	// Simple endpoint to get CUDA info
 	router.GET("/cuda", func(c *gin.Context) {
 		c.JSON(http.StatusOK, gin.H{
 			"cuda_available": cudaAvailable,
@@ -52,39 +57,41 @@ func main() {
 		})
 	})
 
-	// If CUDA is available, add more detailed device info
 	if cudaAvailable {
 		router.GET("/devices", func(c *gin.Context) {
 			devices := make([]string, deviceCount)
 			for i := 0; i < deviceCount; i++ {
 				devices[i] = binding.GetDeviceInfo(i)
 			}
-			c.JSON(http.StatusOK, gin.H{
-				"devices": devices,
-			})
+			c.JSON(http.StatusOK, gin.H{"devices": devices})
 		})
 	}
 
-	// Initialize server
-	serverPort := "8080"
-	if port := os.Getenv("PORT"); port != "" {
-		serverPort = port
+	// Start ngrok listener
+	listener, err := ngrok.Listen(ctx,
+		config.HTTPEndpoint(),
+		ngrok.WithAuthtokenFromEnv(),
+	)
+	if err != nil {
+		return err
 	}
 
-	// Start server in goroutine
+	log.Println("App URL:", listener.URL())
+
+	// Run the server in a goroutine
+	server := &http.Server{Handler: router}
 	go func() {
-		if err := router.Run(":" + serverPort); err != nil {
+		if err := server.Serve(listener); err != nil && err != http.ErrServerClosed {
 			log.Fatalf("Failed to start server: %v", err)
 		}
 	}()
 
-	log.Printf("Server listening on port %s", serverPort)
-
-	// Graceful shutdown
+	// Graceful shutdown handling
 	quit := make(chan os.Signal, 1)
 	signal.Notify(quit, syscall.SIGINT, syscall.SIGTERM)
 	<-quit
 
 	log.Println("Shutting down server...")
-	fmt.Println("Server exited")
+  fmt.Println("Server exited")
+	return server.Shutdown(ctx)
 }
